Add recipe page with favorites and comments

+++ lib/models/recipe_ingredient.dart
@@ -5,7 +5,9 @@ class RecipeIngredient {
   const RecipeIngredient({required this.count, required this.ingredient});
 
   /// Количество в единицах измерения ингредиента
-  final int count;
+  ///
+  /// Дробные количества кратны четверти единицы, например `1.5` или `0.75`
+  final double count;
 
   /// Ингредиент из каталога
   final Ingredient ingredient;
